+++ src/main/java/ch/fhnw/therewrite/controller/DocumentController.java
@@ -1,14 +1,14 @@
 package ch.fhnw.therewrite.controller;
 
-import ch.fhnw.therewrite.data.Annotation;
+import ch.fhnw.therewrite.AccessHelper;
 import ch.fhnw.therewrite.data.Document;
-import ch.fhnw.therewrite.data.Guest;
 import ch.fhnw.therewrite.data.User;
 import ch.fhnw.therewrite.repository.DocumentRepository;
 import ch.fhnw.therewrite.repository.GuestRepository;
 import ch.fhnw.therewrite.repository.UserRepository;
 import ch.fhnw.therewrite.storage.StorageFileNotFoundException;
 import ch.fhnw.therewrite.storage.StorageService;
+import jakarta.servlet.http.HttpSession;
 import org.apache.commons.io.FilenameUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
@@ -16,9 +16,12 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.mock.web.MockMultipartFile;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Paths;
@@ -33,18 +36,28 @@ import java.util.UUID;
 public class DocumentController {
     private final DocumentRepository documentRepository;
     private final StorageService storageService;
-
+    private final UserRepository userRepository;
+    private final GuestRepository guestRepository;
     @Autowired
-    public DocumentController(DocumentRepository documentRepository, StorageService storageService) {
+    public DocumentController(DocumentRepository documentRepository, StorageService storageService, UserRepository userRepository, GuestRepository guestRepository) {
         this.documentRepository = documentRepository;
         this.storageService = storageService;
+        this.userRepository = userRepository;
+        this.guestRepository = guestRepository;
     }
 
+    @PreAuthorize("hasRole('ROLE_USER') or hasRole('ROLE_GUEST')")
     @GetMapping(
             value = "/{documentId}",
             produces = MediaType.APPLICATION_PDF_VALUE
     )
-    public ResponseEntity<byte[]> getDocument(@PathVariable String documentId) {
+    public ResponseEntity<byte[]> getDocument(@PathVariable String documentId, @AuthenticationPrincipal UserDetails currentUser, HttpSession session) {
+        UUID guestId = (UUID) session.getAttribute("guestId");
+        boolean unauthUser = currentUser == null || !AccessHelper.verifyUserRights(currentUser.getUsername(), documentId, documentRepository);
+        boolean unauthGuest = guestId == null || !AccessHelper.verifyGuest(guestId.toString(), documentId, documentRepository, guestRepository);
+        if(unauthUser && unauthGuest) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+        }
         UUID dId = UUID.fromString(documentId);
         Optional<Document> optionalDocument = documentRepository.findById(dId);
         if(optionalDocument.isPresent()) {
@@ -66,6 +79,7 @@ public class DocumentController {
         return documentRepository.findAll();
     }
 
+    @PreAuthorize("hasRole('ROLE_USER')")
     @GetMapping("/all/{userId}")
     public ResponseEntity<List<Document>> getDocumentListByUser(@PathVariable String userId) {
         /*System.out.println(userId);
@@ -79,8 +93,19 @@ public class DocumentController {
         return null;
     }
 
+    @GetMapping("/get/{documentId}")
+    public ResponseEntity<Document> getDocumentById(@PathVariable String documentId) {
+        UUID dId = UUID.fromString(documentId);
+        Optional<Document> optionalDocument = documentRepository.findById(dId);
+        if(optionalDocument.isPresent()) {
+            return ResponseEntity.status(HttpStatus.OK).body(optionalDocument.get());
+        }
+        return ResponseEntity.status(HttpStatus.EXPECTATION_FAILED).body(null);
+    }
+
+    @PreAuthorize("hasRole('ROLE_USER')") // TODO: change to role admin
     @PostMapping("")
-    public ResponseEntity<Document> saveDocument(@RequestParam("file") MultipartFile file, RedirectAttributes redirectAttributes) {
+    public ResponseEntity<Document> saveDocument(@RequestParam("file") MultipartFile file, @AuthenticationPrincipal UserDetails currentUser) {
         String extension = FilenameUtils.getExtension(file.getOriginalFilename());
         if(!Objects.equals(extension, "pdf")){
             return ResponseEntity.status(HttpStatus.UNSUPPORTED_MEDIA_TYPE).body(null);
@@ -93,7 +118,12 @@ public class DocumentController {
             String fileName = document.getId().toString() + ".pdf";
             String filePath = Paths.get(fileName).toString();
             document.setPath(filePath);
-            documentRepository.save(document);
+            List<User> users = document.getUsers();
+            User user = userRepository.findByUsername(currentUser.getUsername());
+            users.add(user);
+            document.setUserCreator(user);
+            document.setUsers(users);
+            Document resp = documentRepository.save(document);
             try {
                 MultipartFile storeFile = new MockMultipartFile(fileName,
                         fileName,
@@ -105,12 +135,29 @@ public class DocumentController {
             catch(IOException exception) {
                 // TODO: log
             }
-            return ResponseEntity.status(HttpStatus.OK).body(document);
+            resp.setUserCreator(user);
+            return ResponseEntity.status(HttpStatus.OK).body(resp);
         }
     }
 
+    @PatchMapping("/{documentId}")
+    public ResponseEntity<Document> updateDocument(@PathVariable UUID documentId, @RequestBody Document updateDetails) {
+        return documentRepository.findById(documentId).map(document -> {
+            document.setSource(updateDetails.getSource());
+            document.setCopyRight(updateDetails.getCopyRight());
+            documentRepository.save(document);
+            return ResponseEntity.ok(document);
+        }).orElse(ResponseEntity.notFound().build());
+    }
+
+
+
+    @PreAuthorize("hasRole('ROLE_USER')") // TODO: change to admin
     @DeleteMapping("/{documentId}")
-    public void deleteDocument(@PathVariable(value="documentId") String documentId) {
+    public void deleteDocument(@PathVariable(value="documentId") String documentId, @AuthenticationPrincipal UserDetails currentUser) {
+        if(currentUser == null) {
+            return;
+        }
         UUID dId = UUID.fromString(documentId);
         documentRepository.deleteById(dId);
     }
